finish functions to get coordinates from position
I updated the place_image() function to take the position as a
parameter, and then figure out where the image should be placed
from that.  It figures this out with get_x/y_from_position functions.
These check if the position is 8 (the center), and then calculate
the coordinates if it isn't.