fix winning line so it draws on top of final move
I was having some issues with getting the winning line drawn
correctly. The place_image() function uses the `onload`
function, which doesn't wait for the image to be loaded before
continuing execution.  Apparently I don't need to use `onload`
at all though - it works fine if I just specify the path to the
image in the `init()` function.