remove event listener for clicks in game_over()
Before, I was determining the game was over by either a winning move,
or by counting the number of x's on the board.  I've changed this so
the handle_mouse_click() function checks for these conditions after
each click and calls game_over() when either the board is full or the
computer has won.